Yet more code cleaning to remove compiler warnings.
svn path=/trunk/; revision=11995

+++ bindings/tk/plr.c
@@ -47,10 +47,14 @@
 
 // Some wrapper macros to return (-1) on error
 
+// Note we use %lu and an explicit cast to unsigned long to print size_t pointers.
+// C99 adds %zd as an explicit format specifier for size_t but this is not yet 
+// fully adopted. 
+
 #define plr_rd( code )                                                            \
     if ( code ) { fprintf( stderr,                                                \
-                      "Unable to read from %s in %s at line %d, bytecount %ld\n", \
-                      plr->iodev->typeName, __FILE__, __LINE__, plr->pdfs->bp );  \
+                      "Unable to read from %s in %s at line %d, bytecount %lu\n", \
+                      plr->iodev->typeName, __FILE__, __LINE__, (unsigned long) plr->pdfs->bp );  \
                   return -1; }
 
 #define plr_cmd( code ) \
@@ -59,8 +63,8 @@
 // Error termination
 
 #define barf( msg )                                               \
-    { fprintf( stderr, "%s\nCommand code: %d, byte count: %ld\n", \
-          msg, csave, plr->pdfs->bp ); return -1; }
+    { fprintf( stderr, "%s\nCommand code: %d, byte count: %lu\n", \
+          msg, csave, (unsigned long) plr->pdfs->bp ); return -1; }
 
 // Static function prototypes.
